requirements-deploy.txt
numpy
ppqi
pillow
loguru
fastapi[all]
paddlepaddle
